feat(modularization): S-10 -- extract skills + messages frontend to separate modules
Two modules (clean separation -- different dep sets, no shared state):

web/modules/skills.js (~647 lines):
  - skills modal (loadSkills, agentApiName from agents.js) +
    skills page (loadGlobalSkills, renderSkillsSidebar, etc.)
  - DI: openModal/closeModal via initSkills({...})
  - clearSkillModalScope() exported for app.js closeModal handler
  - Imports: showToast, t, agentApiName

web/modules/messages.js (~678 lines):
  - Team inter-agent message log + compose + full chat/conversation UI
  - chatSelectedAgent module-private: getChatSelectedAgent()/setChatSelectedAgent()
    exported so app.js federation page and initAgents DI can reach it
  - Imports: showToast, t, avatarBust (from agents.js)
  - No openModal/closeModal needed (no modal in this section)

app.js:
  - Removed three blocks: skills modal (397-583), skills page (1110-1549),
    team+messages (1551-2214) -- 7333 -> 6039 lines
  - closeModal: skillModalScope = null -> clearSkillModalScope()
  - Federation page: direct chatSelectedAgent refs -> getChatSelectedAgent()/
    setChatSelectedAgent()
  - initAgents: setChatSelectedAgent passed directly (no wrapper lambda)

Tests updated to read from the correct modules:
  - messages-view-display-name.test.ts -> messages.js
  - skills-local-on-page.test.ts -> skills.js
  - federation-ui-contract.test.ts -> MESSAGES_MOD for buildBubbleHtml test

Verified: node --check both modules, npm run build, 2873/2873 vitest passing,
Playwright headless smoke 0 JS errors (skills/messages/connectors/memories/kanban).
